Move shared type definitions to top-level `types` directory
