--- packages/apollo-server-core/src/types/mock-req.d.ts
@@ -1,12 +1,0 @@
-declare module 'mock-req' {
-  import { Request, Headers } from 'apollo-server-env';
-
-  class MockReq implements Pick<Request, 'method' | 'url' | 'headers'> {
-    constructor();
-    method: string;
-    url: string;
-    headers: Headers;
-  }
-
-  export = MockReq;
-}
