+++ tsconfig.base.json
@@ -16,6 +16,10 @@
     "noUnusedParameters": true,
     "noUnusedLocals": true,
     "lib": ["es2017", "esnext.asynciterable"],
-    "types": ["node", "jest"]
+    "types": ["node", "jest"],
+    "baseUrl": ".",
+    "paths": {
+      "*" : ["types/*"]
+    }
   }
 }
