+++ types/mock-req/index.d.ts
@@ -1,0 +1,10 @@
+import { Request, Headers } from 'apollo-server-env';
+
+declare class MockReq implements Pick<Request, 'method' | 'url' | 'headers'> {
+  constructor();
+  method: string;
+  url: string;
+  headers: Headers;
+}
+
+export = MockReq;
